config.py
import mysql.connector

def get_db():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password="Shrikant",
        database="realestate_platform"
    )